+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## [2.0.0]
+- Max deep user settings
+- Only show container objects if has children
+- Link objects config
+
 ## [1.0.0]
 - First version